+++ kingfisher_scrapy/spiders/nigeria_kaduna_state_records.py
@@ -1,0 +1,20 @@
+from kingfisher_scrapy.spiders.nigeria_kaduna_state_base import NigeriaKadunaStateBudeshiBase
+from kingfisher_scrapy.util import components
+
+
+class KadunaStateBudeshiRecords(NigeriaKadunaStateBudeshiBase):
+    """
+    Domain
+        Kaduna State
+    Bulk download documentation
+        https://www.budeshi.ng/kadppa/api
+    """
+    name = 'nigeria_kaduna_state_records'
+
+    # SimpleSpider
+    data_type = 'record_package'
+
+    def build_urls(self, project):
+        id = project['id']
+        url = f'{self.base_url}record/{id}'
+        yield self.build_request(url, formatter=components(-2))
